Created a serverless stlite version
Several challenges were encountered in development of this [stlite](https://github.com/whitphx/stlite) app. Hopefully, documentation of them will help the unwary traveler navigate them faster.

- stlite is built on [pyodide](https://pyodide.org/). Pyodide is under rapid development and stlite currently supports version 0.27.2. There are three ways you can use a Python package.  (1) It is a pure Python wheel, (2) it has been built in to Pyodide, or (3) it is built correctly outside of pyodide and loaded directly (e.g., polars). See https://pyodide.org/en/0.27.2/usage/loading-packages.html for details.
- Getting [polars](https://pypi.org/project/polars/) to work was a bear (haha).  It is supposedly included in pyodide including the version that was currently supported by stlite (see https://pyodide.org/en/0.27.2/usage/packages-in-pyodide.html#packages-in-pyodide) but it didn't load by the normal mechanisms. Internet searches left it a bit unclear whether polars was supported by pyodide yet, but it turned out that polars has been building a wheel that supports pyodide and including it in its more recent releases.  The most recent version uses an [enscripten](https://emscripten.org/) version that was incompatible with what pyodide was built in.  So needed to go to version 1.20 of polars.  The wheel can be found in the assets of the release (see https://github.com/pola-rs/polars/releases/tag/py-1.20.0 specifically https://github.com/pola-rs/polars/releases/download/py-1.20.0/polars-1.20.0-cp39-abi3-emscripten_3_1_58_wasm32.whl).  This is then included directly and installed.
- Next was getting [griddler](https://github.com/CDCgov/pygriddler) to work. Because it uses progressbar a pure Python wheel could not be created. I patched it to remove the code using progressbar and created a wheel.
- I made a wheel of the metapop package that does the simulation and creates the steamlit app.  For some reason, the wheel made by poetry was not recognized as a correctly formatted wheel by [micropip](https://pyodide.org/en/stable/usage/loading-packages.html#micropip).  Suspect there is a bug there either in poetry or micropip, but I'm not sure which and what is causing it as griddler's wheel built fine with poetry.  So I switched the build system to hatchling
- In `st.dataframe` was getting pyarrow failures due to the conversion of the polars data to pandas.  It said it was version 0.0.1.  I was very confused... I later figured out that [stlite was mocking pyarrow](https://github.com/whitphx/stlite/issues/656) to avoid problems because it wasn't yet supported by pyodide.  But it is since... so I was able to remove the mocked version (thank goodness) and install it directly.  This will be fixed in a future version of stlite.
- However, `st.dataframe` in our app was returning a blank box where the table should (but I could download the CSV!).  The reasons are unclear as I tested the same table in a smaller steamlit app and it displayed fine.  This seems to likely be a bug in stlite but I've not been able to isolate it.  I replaced it with a markdown version of the table and a [data URI](https://en.wikipedia.org/wiki/Data_URI_scheme) I generated from a CSV internally so that folks could still download the CSV only when pyodide is loaded.
- With all of the above it works if you have a server to serve the wheels and python files that are referenced in the HTML, but we wanted a single page app.  So, I wrote a build script `stlite/make_measles_sim_html` that replaced the URLs with a zip archive that was written as a Javascript UInt8 array and imported via the [loading archive files mechanism in stlite](https://github.com/whitphx/stlite#loading-archive-files-archives-option).

+++ metapop/app_helper.py
@@ -1279,7 +1279,9 @@ def get_table(combined_results, IHR):
     combined_results = combined_results.with_columns(
         pl.Series(
             name="Hospitalizations",
-            values=np.random.binomial(combined_results["Total"].to_numpy(), IHR),
+            values=np.random.binomial(
+                combined_results["Total"].to_numpy().astype("int32"), IHR
+            ),
         )
     )
